+++ src/test/java/fr/lacombe/BankAcceptanceTest.java
@@ -3,6 +3,8 @@ package fr.lacombe;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 
+import java.math.BigDecimal;
+
 public class BankAcceptanceTest {
     @Test
     public void deposit_and_withdrawal_on_account_print_the_statement() {
@@ -11,9 +13,9 @@ public class BankAcceptanceTest {
         OperationsHistory operationsHistory = new OperationsHistory();
         ClockFake clockOperation = new ClockFake("10/01/2012", "13/01/2012", "14/01/2012");
         Account account = new Account(printer, operationsHistory, new FormatterOperation(), clockOperation);
-        account.deposit(new Amount(1000.00));
-        account.deposit(new Amount(2000.00));
-        account.withdrawal(new Amount(500.00));
+        account.deposit(new Amount(BigDecimal.valueOf(1000.00)));
+        account.deposit(new Amount(BigDecimal.valueOf(2000.00)));
+        account.withdrawal(new Amount(BigDecimal.valueOf(500.00)));
 
         //test
         account.print();
